feat(venue): remove VenueBody old
